added submit button and used code from practice and flowlayout
